+++ lib/models/room.js
@@ -1,0 +1,14 @@
+"use strict";
+
+function Room(roomId) {
+    this.roomId = roomId;
+    this.name = roomId;
+    this.timeline = [];
+    this.oldState = null;
+    this.currentState = null;
+}
+
+/**
+ * The Room class.
+ */
+module.exports = Room;
